requirements.txt
Flask==3.1.0
python-dotenv>=0.19.0
requests==2.25.1
gunicorn==23.0.0
mercadopago>=2.0.0
psycopg2-binary>=2.9.1